pyproject.toml
[project]
name = "graphql-api-test"
version = "0.1.0"
description = ""
requires-python = ">=3.14"
dependencies = [
    "faker==40.4.0",
    "requests==2.32.5",
    "pytest==9.0.2",
    "pytest-xdist==3.8.0",
    "allure-pytest==2.15.3",
    "python-dotenv==1.2.1",
    "pydantic==2.12.5",
]


[dependency-groups]
dev = [
    "ruff>=0.15.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/api"]

[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_functions = ["test_*"]
norecursedirs = [".git", ".venv", ".idea", "__pycache__", "build", "dist"]
addopts = [
    "-v",
    "-ra",
    "--strict-markers",
    "--tb=short",
    "--alluredir=test-result/allure-results",
]
markers = [
    "smoke: Quick smoke tests for basic functionality",
    "regression: Full regression test suite",
]
console_output_style = "progress"
log_cli = true
log_cli_level = "INFO"
log_cli_format = "%(asctime)s [%(levelname)s] %(message)s"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"
xfail_strict = true

[tool.ruff]
target-version = "py314"
line-length = 120
exclude = [
    ".git",
    ".venv",
    "__pycache__",
    "test-result",
    "pytest-cache-files-*",
]

[tool.ruff.lint]
select = [
    "E", # pycodestyle errors
    "W", # pycodestyle warnings
    "F", # Pyflakes
    "I", # isort
    "B", # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
    "SIM", # flake8-simplify
    "RUF", # Ruff-specific rules
]
ignore = [
    "E501", # line too long (handled by formatter)
    "B008", # do not perform function calls in argument defaults (needed for fixtures)
    "RUF013", # allow implicit Optional in existing API wrappers
    "UP042", # keep legacy `str, Enum` declarations in generated models
]
fixable = ["ALL"]
unfixable = []

[tool.ruff.lint.isort]
known-first-party = ["src", "tests"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = [
    "B011", # allow assert in tests
    "F821", # allow forward/missing typing names in generated tests
    "F841", # allow intermediate variables in tests
    "SIM101", # keep explicit assertions style
    "SIM105", # allow try/except cleanup in tests
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
